requirements.txt
streamlit
langchain
langchain-community
pypdf
python-dotenv
faiss-cpu
sentence-transformers
ollama
langchain-ollama